Implementat iteratia 2.

+++ src/main/java/repository/ILocRepository.java
@@ -2,5 +2,5 @@ package repository;
 
 import model.Loc;
 
-public interface ILocRepository extends ICRUDRepository<Long, Loc> {
+public interface ILocRepository extends ICRUDRepository<Integer, Loc> {
 }
