+++ src/main/java/repository/ISpectatorRepository.java
@@ -2,5 +2,5 @@ package repository;
 
 import model.Spectator;
 
-public interface ISpectatorRepository extends ICRUDRepository<Long, Spectator> {
+public interface ISpectatorRepository extends ICRUDRepository<Integer, Spectator> {
 }
